Quiz example (#3)
* Implemented the quiz.
Most of the quiz is completed. Randomization was not directly implemented. Instead the map data structure was used with the range operator, which results in random access of the map elements (the ordering of Go maps is not defined for integers?).
I'll try to implement a better random sequence of questions later.

* Implemented random question generation switched by command line flag -random.
This completely implements the quiz according to the specification.