Removing a broken link in the README
The README has a broken link to the `problem.csv` file
so I removed that link and instead just noted that
an example of that file is shown below. This really
shouldn't affect the overall readability of the problem
description, but if anyone sees an issue with it let
me know.

I also added a note that mentions a possible CSV row
that often will break custom CSV parsers to encourage
devs to look into the `encoding/csv` package in Go.

Fixes #20